Support 2014 and 2015 clients (thanks to rAthena).

+++ src/map/buyingstore.h
@@ -26,7 +26,7 @@ DBMap *buyingstore_db; // Db holder the buyer : charid -> map_session_data
 
 // Autotrade entry
 struct s_autotrade_entry {
-	uint16 cartinventory_id; // Item entry id/cartinventory_id in cart_inventory table (for vending)
+	uint32 cartinventory_id; // Item entry id/cartinventory_id in cart_inventory table (for vending)
 	uint16 amount; // Amount
 	uint32 price; // Price
 	uint16 index; // Item index in cart
@@ -35,7 +35,7 @@ struct s_autotrade_entry {
 
 // Struct of autotrader
 struct s_autotrader {
-	uint16 id; // vendor/buyer id
+	uint32 id; // vendor/buyer id
 	uint32 account_id; // Account ID
 	uint32 char_id; // Char ID
 	int m; // Map location
